requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
motor>=3.1.1
python-dotenv>=1.0.0
dnspython>=2.3.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
